+++ CHANGES.md
@@ -2,6 +2,6 @@
 
 - Based on squoosh-multiple-export by Khongchai (https://github.com/Khongchai/squoosh-multiple-export), forked from Squoosh by Google Chrome Labs (https://github.com/GoogleChromeLabs/squoosh).
 - Forked and renamed to Squiish.
-- Bulk export workflow updates (batch size, sequential rename/prepend).
+- Bulk export workflow updates (batch size, sequential rename/append).
 - Resize options updated (long edge mode, UI adjustments).
 - Telemetry removed.